backend/requirements.txt
annotated-types==0.7.0
anthropic==0.43.0
anyio==4.8.0
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
nltk==3.9.1
openai==1.59.7
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
